Add laptop assignee field
A laptop has an owner, which continues to hold the concept of the
person who has physical posession of the laptop in this specific moment.

A laptop now additionally has an assignee, who is the logical owner
of the laptop - the person who should have the laptop if the laptop
is not in the process of delivery or repair. (i.e. the student, or in
the case of Peru where students are not tracked, the school director)