Add form for handout by classroom
The "Register handout" functionality now primarily shows a list of
students (by class) who have laptops assigned but not in their hands.

Untick the ones who didn't receive laptops in the handout, then click Save.

You can still input a list of serial numbers too, but that is expected
to be the corner case.