Abm2: reset to first page on new search
Fix a bug that can be reproduced by:

 1. Open Laptops controller with a lot of laptops
 2. Go to last page of results
 3. Search for a specific serial number

No laptop is displayed, and it tells you that you are on page 26 of 1.
Returning back to the first page brings up the result, but this is a bit
non-obvious.

Fix this by resetting to the first page when performing a new search.