ABM: Communicate sort choices to database
Right now, the table model is only ever populated with one page of items.
Sorting the columns sorts those items, without consiering the other items
in the database.

Trick qooxdoo into not actually performing a sort on the data (but make
the UI retain the suggestion of sorting), and communicate sort changes
to the database, updating the model appropriately each time.