Add config mechanism, allow customisation of page title
Add a YAML-based configuration mechanism. The idea is that all values
will be optional (and you don't need to put a config file at all
if you don't wish to customise anything).

As the first option, allow customisation of the page title. This
involved a bit of reshuffling of root routes and index.html pages.