Extend "Move people" functionality
In Nicaragua, the team is often faced with the problem that at the
end of the school year, Segundo A and Segundo B gets thoroughly mixed up,
and the following year's Tercero A and Tercero B bear little or no
relation to the previous A/B section split from the previous year.

This patch introduces a tool to quickly and easily move a significant
number of people from one place to another, replacing the old
"Transfer people" functionality which always moved everyone.
The user selects the source and destination places, and is then
presented with a list of people to move which can be ticked/unticked.
Finally, all of the ticked people are moved to the destination.