Reorganise gui integration in codebase
Integrate gui closer with the ruby codebase, based on ideas from
http://qooxdoo.org/documentation/general/ruby_on_rails_integration

This means that we can default to loading the source directly in the
development environment (after generate.py source) which is much
more convenient and allows for debugging.

compile_gui.sh was removed and replaced by a rake task e.g.
    rake gui:generate (for source)
    rake gui:generate[build] (for build or other targets)

The production environment will look for a built GUI so things must
be built as before.