Barcode scan form: improve design and behaviour
The barcode scan form presented a combo-box of locations to select from.
This is proving hard work now that we have over 50 schools, each with
a number of turns and sections.

Move to a wizard-like approach where the location is selected on the first
screen from a tree.

Previously, mass assignments/movements of teachers laptops was difficult
as the previous implementation avoided anything but transferring to
students in classes. Make the system more general so that we can include
teachers here: we recently had to assign/move laptops to hundreds of
new Nicaraguan teachers in the project, and it was a pain.