GUI: Improve linkage between laptop and owner
There is currently no immediate/direct way of searching for a laptop
serial number and finding out the owner's location.

Fix this by creating a hyperlink in the Laptop details view which
links to the person information (where the location information can
be quickly seen).

Also remove the ability to directly edit a laptop's owner/assignee
(this should be done with movements or assignments so that history
is correctly recorded) as well as the ability to specify an assignee
when creating laptops (this should be done with the assignments
system).

Requested by the Nicaraguan inventory team.