Dispose all elements on logout
As suspected by the comment in this code, we were not disposing the UI
elements when logging out. Fix this by disposing children as done in
other places.

This requires us to turn the taskbar and logout button widgets into
regular, dynamic classes - there was not much reason for them to be
singleton.