Barcode report: rework and extend
Move the barcode report configuration dialog into its own class, so
that we can created a more complicated UI.

Add the ability to produce more detailed box labels, including names of
laptops to be included inside.

Add the ability to produce laptop name labels, intended to be stuck
on the handle of the XO.

Improve the data structure passed between model and view so that
the students passed are grouped by the box they are suggested to be
placed inside, which simplifies the view code.

This new functionality was requested by the Nicaragua inventario team.