Improve report generation
Reports are now sent to the client as new windows, which get presented
by modern browsers as downloads. Avoids the iframe hack and is much
simpler.

Added the option to export all PDF reports as HTML using the same
mechanism.

Combined the "People and their laptops" UI for PDF and XLS into one,
which now exports as PDF, XLS or HTML.