Place.getSerialsInfo: search in ancestor places too
Clarified by Martin Abente. It is necessary to include ancestor
places in the search for laptops (when deciding which laptops to
generate activations for for a specific place), because in Paraguay,
certain formadores are registered in the city place because they require
the ability to activate their laptops in any school in their district.