Allow lease expiry date to be specified
In Nicaragua there are some cases where we want leases to expiry on a
certain day. Add a field where this can be specified; the expiry/duration
fields must now be set on an either-or basis.

Also modify the lease duration column to be in units of days, which
is more sensible.