Connection events functionality
Add functionality to receive per-laptop "connection events" over HTTP.
These will be saved in the database along with certain bits of metadata
provided in the submission.

It is compatible with a new mechanism added to oatslite for submitting
this.

This will allow a few items of interest to be monitored:
 1. How many laptops are coming online every day
 2. How much free disk space those laptops have
    (to see whether they could complete a software update)
 3. What software version they are running
    (to see whether an automatic software update completed)