Refine lease generation for laptops in ancestor places
Having recollected laptops in for the school holidays here, there are
many laptops in a parent location, which are "in use" (since they will
be returned to schools soon this is arguably correct), and assigned to
a student in some child location.

But this matches our criteria for ancestor places for all of the schools,
so we have too many leases being sent to too many places.

Now we only include laptops in ancestor places where the laptop is
both "in use" and unassigned (quite a strange case which should not
extend beyond the "formador" use case), and "in use" and assigned to the
same person (another version of the "formador" use case).