Lease info must now be requested per-school
schools_leases was taking so long on the Nica production db that
it was causing timeouts. There are too many schools now to be able
to do this all at once - that was never sensible.

Move to a model where each school must be requested individually.

Also remove some old/unused REST methods from the places controller.