Cleaner column specifications and SC/ABM cleanup
Cleaned up SearchController/ABM and related things.

Models now specify column specifications in a more simplistic manner.
Normally they just provide the column name, and if it is a foreign column,
the association which should be traversed to find the column.

In the case where an alternative display is wanted (or there is no
directly corresponding DB column that can be included in the query),
they can provide their own attribute (method) to be called.

Thats all that must be provided. The rest (figuring out sort/search
columns, etc) is auto-detected.